Add class to get map object
