+++ docs/style.css
@@ -4,7 +4,7 @@ body {
   margin-right: auto;
 }
 
-.tilecloud {
+.tilecloud, #my-map {
   width: 100%;
   height: 400px;
 }
