+++ src/embed.js
@@ -4,16 +4,16 @@
 
 import '@babel/polyfill'
 import 'intersection-observer'
+import mapboxgl from 'mapbox-gl'
 import 'mapbox-gl/dist/mapbox-gl.css'
-import render from './lib/render'
+import TilecloudMap from './lib/tilecloud-map'
 
 const observer = new IntersectionObserver(entries => {
   entries.forEach(item => {
     if (!item.isIntersecting) {
       return
     }
-
-    render(item.target)
+    new TilecloudMap(item.target)
     observer.unobserve(item.target)
   })
 })
@@ -23,3 +23,8 @@ const containers = document.querySelectorAll('.tilecloud')
 containers.forEach(container => {
   observer.observe(container)
 })
+
+window.tilecloud = {
+  Map: TilecloudMap,
+  mapboxgl,
+}
